+++ core/src/main/thrift/tabletmgmt.thrift
@@ -72,14 +72,4 @@ service TabletManagementClientService {
     3:string lock
     4:data.TKeyExtent extent
   )
-  
-  void requestTabletHosting(
-    1:client.TInfo tinfo
-    2:security.TCredentials credentials
-    3:string tableId
-    4:list<data.TKeyExtent> extents
-  ) throws (
-    1:client.ThriftSecurityException sec
-  )
-
 }
