+++ server/manager/src/main/java/org/apache/accumulo/manager/split/Splitter.java
@@ -100,7 +100,7 @@ public class Splitter {
 
   LoadingCache<CacheKey,FileInfo> splitFileCache;
 
-  private static int weigh(KeyExtent keyExtent) {
+  public static int weigh(KeyExtent keyExtent) {
     int size = 0;
     size += keyExtent.tableId().toString().length();
     if (keyExtent.endRow() != null) {
